--- src/nhs_reusable_code_library/standard_data_validations/__init__.py
@@ -1,20 +1,0 @@
-def mod11_check(nhsNumber :str) -> bool:
-    """This function checks if the NHS number passes the MOD 11 check.
-    Args:
-    nhsNumber (str): The NHS number to be checked.
-    Returns:
-    bool: True if the NHS number passes the MOD 11 check, False otherwise.
-    """
-    pass
-
-
-def sensitive_legally_restricted(decision_to_admit :str, nhs_number_status_indicator :str, witheld_identity_reason :str) -> bool:
-    """ This function checks if the NHS number is sensitive or legally restricted.
-    Args:
-    decision_to_admit (str): The decision to admit the patient.
-    nhs_number_status_indicator (str): The NHS number status indicator.
-    witheld_identity_reason (str): The reason for withholding the identity.
-    Returns:
-    bool: True if the NHS number is sensitive or legally restricted, False otherwise.
-    """
-    pass
